biddingapi/requirements.txt
amqp==5.0.6
APScheduler==3.8.1
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.1
celery-once==3.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
h11==0.12.0
idna==3.3
kombu==5.2.2
lxml==4.6.4
openpyxl==3.0.9
outcome==1.1.0
Pillow==8.4.0
prompt-toolkit==3.0.22
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.0.2
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
